OS Bot/requirements.txt
discord.py==2.1.1
openai==0.26.5
requests==2.28.2
youtube_dl==2021.12.17
